examples/pydantic_ai/requirements.txt
autotel>=0.1.0
pydantic-ai>=1.19.0
ollama>=0.6.0
